Fixes #15 begin restructuring
