+++ .gitignore
@@ -167,3 +167,19 @@ app/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-11-01-19.290950.txt
 app/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-11-01-19.503072.txt
 app/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-11-02-17.813932.txt
 app/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-11-02-18.020168.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-06-20-24-52.117075.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-06-20-25-36.514458.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-10-54-57.536320.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-10-54-57.750310.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-10-57-26.621275.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-10-57-26.831587.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-11-01-19.290950.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-11-01-19.503072.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-11-02-17.813932.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_25-vns-max-2023-09-09-11-02-18.020168.txt
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_77-vns-max-2023-07-09-14-57-21.147373.csv
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_77-vns-max-2023-07-09-14-59-24.220923.csv
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_77-vns-max-2023-07-09-18-10-10.370226.csv
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_77-vns-max-2023-07-09-18-10-55.459532.csv
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_77-vns-max-2023-07-10-14-01-40.827858.csv
+app/single_objective_oprimization_problems/trivial/max_ones_problem/outputs/dimension_77-vns-max-2023-07-10-17-15-25.484215.csv
