+++ app/single_objective_oprimization_problems/trivial/max_ones_problem/command_line.py
@@ -7,6 +7,8 @@ directory = Path(__file__).resolve()
 sys.path.append(directory.parent)
 sys.path.append(directory.parent.parent)
 sys.path.append(directory.parent.parent.parent)
+sys.path.append(directory.parent.parent.parent.parent)
+sys.path.append(directory.parent.parent.parent.parent.parent)
 
 import os
 import logging
