+++ app/single_objective_oprimization_problems/trivial/max_ones_problem/inputs/dimension_25.txt
