+++ app/single_objective_oprimization_problems/trivial/max_ones_problem/max_ones_problem.py
@@ -9,6 +9,8 @@ directory = Path(__file__).resolve()
 sys.path.append(directory.parent)
 sys.path.append(directory.parent.parent)
 sys.path.append(directory.parent.parent.parent)
+sys.path.append(directory.parent.parent.parent.parent)
+sys.path.append(directory.parent.parent.parent.parent.parent)
 
 from copy import deepcopy
