+++ app/single_objective_oprimization_problems/trivial/max_ones_problem/max_ones_problem_binary_bit_array_solution_vns_support.py
@@ -1,36 +1,53 @@
 """ 
-.. _py_max_ones_problem_int_solution_vns_support:
+..  _py_max_ones_problem_bit_array_solution_vns_support:
 
-The :mod:`~app.max_ones_problem.max_ones_problem_binary_int_solution_vns_support` contains class :class:`~app.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport`, that represents solution of the :ref:`Problem_Max_Ones`, where `int` representation of the problem has been used.
+The :mod:`~app.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support` contains class :class:`~app.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport`, that represents supporting parts of the `VNS` algorithm, where solution of the :ref:`Problem_Max_Ones` have `BitArray` representation.
 """
 
 import sys
 from pathlib import Path
 directory = Path(__file__).resolve()
+sys.path.append(directory)
 sys.path.append(directory.parent)
-sys.path.append(directory.parent.parent)
 sys.path.append(directory.parent.parent.parent)
+sys.path.append(directory.parent.parent.parent.parent)
+sys.path.append(directory.parent.parent.parent.parent.parent)
 
 from copy import deepcopy
-
 from random import choice
-from random import randint
+from random import random
+from bitstring import BitArray
 
 from uo.target_problem.target_problem import TargetProblem
 from uo.target_solution.target_solution import ObjectiveFitnessFeasibility
 from uo.target_solution.target_solution import TargetSolution
 from uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support import ProblemSolutionVnsSupport
 
-class MaxOnesProblemBinaryIntSolutionVnsSupport(ProblemSolutionVnsSupport):
+from app.utils.logger import logger
+from app.max_ones_problem.max_ones_problem import MaxOnesProblem
+
+
+class MaxOnesProblemBinaryBitArraySolutionVnsSupport(ProblemSolutionVnsSupport):
     
     def __init__(self)->None:
+        """
+        Create new MaxOnesProblemBinaryBitArraySolutionVnsSupport instance
+        """
         return
 
     def __copy__(self):
+        """
+        Internal copy of the MaxOnesProblemBinaryBitArraySolutionVnsSupport
+        :return: MaxOnesProblemBinaryBitArraySolutionVnsSupport -- new MaxOnesProblemBinaryBitArraySolutionVnsSupport instance with the same properties
+        """
         sol = deepcopy(self)
         return sol
 
     def copy(self):
+        """
+        Copy the MaxOnesProblemBinaryBitArraySolutionVnsSupport
+        :return: MaxOnesProblemBinaryBitArraySolutionVnsSupport -- new MaxOnesProblemBinaryBitArraySolutionVnsSupport instance with the same properties
+        """
         return self.__copy__()
         
     def vns_randomize(self, k:int, problem:TargetProblem, solution:TargetSolution, solution_codes:list[str])->bool:
@@ -51,17 +68,14 @@ class MaxOnesProblemBinaryIntSolutionVnsSupport(ProblemSolutionVnsSupport):
             positions:list[int] = []
             for i in range(0,k):
                 positions.append(choice(range(k)))
-            new_representation:int = solution.representation
-            mask:int = 0
-            for p in positions:
-                mask |= 1 << p
-            mask = ~mask
-            solution.representation ^= mask
+            new_representation:BitArray = deepcopy(solution.representation)
+            new_representation.invert(positions)
             all_ok:bool = True
+            #logger.debug(solution_codes)
             for sc in solution_codes:
-                sc_representation = int(sc,2)
-                if sc_representation != 0:
-                    comp_result:int = (sc_representation ^ new_representation).bit_count()
+                sc_representation = solution.representation_string_to_bit_array(sc)
+                if sc_representation is not None and sc_representation != '':
+                    comp_result:int = (sc_representation ^ new_representation).count(value=1)
                     if comp_result > k:
                         all_ok = False
             if all_ok:
@@ -72,7 +86,7 @@ class MaxOnesProblemBinaryIntSolutionVnsSupport(ProblemSolutionVnsSupport):
             return True
         else:
             return False 
-
+        
     def string_representation(self, delimiter:str, indentation:int=0, indentation_symbol:str='', group_start:str ='{', 
         group_end:str ='}')->str:
         """
@@ -91,7 +105,7 @@ class MaxOnesProblemBinaryIntSolutionVnsSupport(ProblemSolutionVnsSupport):
         :return: string representation of instance that controls output
         :rtype: str
         """        
-        return 'MaxOnesProblemBinaryIntSolutionVnsSupport'
+        return 'MaxOnesProblemBinaryBitArraySolutionVnsSupport'
 
     def __str__(self)->str:
         """
@@ -121,3 +135,5 @@ class MaxOnesProblemBinaryIntSolutionVnsSupport(ProblemSolutionVnsSupport):
         :rtype: str
         """
         return self.string_representation('|')
+
+
