+++ app/single_objective_oprimization_problems/trivial/max_ones_problem/max_ones_problem_binary_int_solution.py
@@ -10,6 +10,8 @@ directory = Path(__file__).resolve()
 sys.path.append(directory.parent)
 sys.path.append(directory.parent.parent)
 sys.path.append(directory.parent.parent.parent)
+sys.path.append(directory.parent.parent.parent.parent)
+sys.path.append(directory.parent.parent.parent.parent.parent)
 
 from copy import deepcopy
 from random import choice
